NetworkService class has been implemented
Unit test for successful scenario written. But test fails with 401.
It is due to NetworkService is really making real network request and fails with 401.
For unit tests, we should not make network requests as it takes relatively more time.
Cumulatively all tests takes lot of time to complete test suite. As a result of that,
it slow down continuous integration.

Next Step:
We need to mock the URLSession with  MockURLSession to stub the mock responses.